Remove has_my_FIN_been_sent and has_my_FIN_been_acked
These are unnecessary. We can determine this state by looking at
has_EOF_been_read as well as the list of unacked segments. If EOF
is read, has_EOF_been_read is set to TRUE and a FIN segment is
added to the end of our list of unacked segments.